fix issue with host url not working
